+++ Scrapers/Data/DK/CBB.json
@@ -1,14 +1,14 @@
 [
     {
         "Away Team": "Clemson",
-        "Away Team Rank Name": "Clemson",
+        "Away Team Rank Name": "11ff4566",
         "DK Away Odds": {
             "Spread": "+7.5",
             "Spread Odds": "−112",
             "Away ML": "+250"
         },
         "Home Team": "Arizona",
-        "Home Team Rank Name": "Arizona",
+        "Home Team Rank Name": "2b1ac347",
         "DK Home Odds": {
             "Spread": "-7.5",
             "Spread Odds": "-7.5",
@@ -16,22 +16,22 @@
         },
         "Game": {
             "Start Time": "7:09PM",
-            "Total": "152.5",
-            "Over Total Odds": "−108",
-            "Under Total Odds": "−112",
-            "League": "NCAAB"
+            "Total": "152",
+            "Over Total Odds": "−112",
+            "Under Total Odds": "−108",
+            "League": "CBB"
         }
     },
     {
         "Away Team": "San Diego State",
-        "Away Team Rank Name": "San Diego St",
+        "Away Team Rank Name": "b1915b99",
         "DK Away Odds": {
             "Spread": "+11",
-            "Spread Odds": "−108",
+            "Spread Odds": "−105",
             "Away ML": "+500"
         },
         "Home Team": "UConn",
-        "Home Team Rank Name": "Connecticut",
+        "Home Team Rank Name": "722d49f",
         "DK Home Odds": {
             "Spread": "-11",
             "Spread Odds": "-11",
@@ -40,24 +40,24 @@
         "Game": {
             "Start Time": "7:39PM",
             "Total": "136",
-            "Over Total Odds": "−112",
-            "Under Total Odds": "−108",
-            "League": "NCAAB"
+            "Over Total Odds": "−110",
+            "Under Total Odds": "−110",
+            "League": "CBB"
         }
     },
     {
         "Away Team": "Alabama",
-        "Away Team Rank Name": "Alabama",
+        "Away Team Rank Name": "96a8ead4",
         "DK Away Odds": {
-            "Spread": "+4.5",
-            "Spread Odds": "−110",
+            "Spread": "+5",
+            "Spread Odds": "−115",
             "Away ML": "+164"
         },
         "Home Team": "North Carolina",
-        "Home Team Rank Name": "N Carolina",
+        "Home Team Rank Name": "c16ebc3c",
         "DK Home Odds": {
-            "Spread": "-4.5",
-            "Spread Odds": "-4.5",
+            "Spread": "-5",
+            "Spread Odds": "-5",
             "Home ML": "−198"
         },
         "Game": {
@@ -65,30 +65,30 @@
             "Total": "173.5",
             "Over Total Odds": "−110",
             "Under Total Odds": "−110",
-            "League": "NCAAB"
+            "League": "CBB"
         }
     },
     {
         "Away Team": "Illinois",
-        "Away Team Rank Name": "Illinois",
+        "Away Team Rank Name": "11b30666",
         "DK Away Odds": {
             "Spread": "+1.5",
-            "Spread Odds": "−112",
-            "Away ML": "+102"
+            "Spread Odds": "−110",
+            "Away ML": "+100"
         },
         "Home Team": "Iowa State",
-        "Home Team Rank Name": "Iowa St",
+        "Home Team Rank Name": "838603b0",
         "DK Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "-1.5",
-            "Home ML": "−122"
+            "Home ML": "−120"
         },
         "Game": {
             "Start Time": "10:20PM",
             "Total": "146.5",
-            "Over Total Odds": "−110",
-            "Under Total Odds": "−110",
-            "League": "NCAAB"
+            "Over Total Odds": "−108",
+            "Under Total Odds": "−112",
+            "League": "CBB"
         }
     }
 ]
