+++ Scrapers/Data/DK/NBA.json
@@ -1,22 +1,22 @@
 [
     {
         "Away Team": "BOS Celtics",
-        "Away Team Rank Name": "Boston",
+        "Away Team Rank Name": "8a2d1e91",
         "DK Away Odds": {
-            "Spread": "-16",
-            "Spread Odds": "−105",
-            "Away ML": "−1800"
+            "Spread": "-16.5",
+            "Spread Odds": "−108",
+            "Away ML": "−1350"
         },
         "Home Team": "ATL Hawks",
-        "Home Team Rank Name": "Atlanta",
+        "Home Team Rank Name": "354ef58",
         "DK Home Odds": {
-            "Spread": "+16",
-            "Spread Odds": "−115",
-            "Home ML": "+1000"
+            "Spread": "+16.5",
+            "Spread Odds": "−112",
+            "Home ML": "+800"
         },
         "Game": {
             "Start Time": "7:40PM",
-            "Total": "227.5",
+            "Total": "226",
             "Over Total Odds": "−110",
             "Under Total Odds": "−110",
             "League": "NBA"
@@ -24,24 +24,24 @@
     },
     {
         "Away Team": "MIL Bucks",
-        "Away Team Rank Name": "Milwaukee",
+        "Away Team Rank Name": "c5227021",
         "DK Away Odds": {
-            "Spread": "-2.5",
-            "Spread Odds": "−105",
-            "Away ML": "−135"
+            "Spread": "-1",
+            "Spread Odds": "−108",
+            "Away ML": "−115"
         },
         "Home Team": "NO Pelicans",
-        "Home Team Rank Name": "New Orleans",
+        "Home Team Rank Name": "6a3773e4",
         "DK Home Odds": {
-            "Spread": "+2.5",
-            "Spread Odds": "−115",
-            "Home ML": "+114"
+            "Spread": "+1",
+            "Spread Odds": "−112",
+            "Home ML": "−105"
         },
         "Game": {
             "Start Time": "8:10PM",
             "Total": "224.5",
-            "Over Total Odds": "−110",
-            "Under Total Odds": "−110",
+            "Over Total Odds": "−108",
+            "Under Total Odds": "−112",
             "League": "NBA"
         }
     }
