+++ Scrapers/Data/ESPN/MLB.json
@@ -40,8 +40,8 @@
         "Game": {
             "Start Time": "Today \u00b7 4:10 PM",
             "Total": "7.5",
-            "Over Total Odds": "-115",
-            "Under Total Odds": "-105",
+            "Over Total Odds": "-110",
+            "Under Total Odds": "-110",
             "League": "MLB"
         }
     },
@@ -119,21 +119,21 @@
         "Away TeamID": "486a5e3b",
         "ESPN Away Odds": {
             "Spread": "-1.5",
-            "Spread Odds": "+160",
+            "Spread Odds": "+150",
             "Away ML": "-110"
         },
         "Home Team": "SD Padres",
         "Home TeamID": "3f4235d4",
         "ESPN Home Odds": {
             "Spread": "+1.5",
-            "Spread Odds": "-190",
+            "Spread Odds": "-180",
             "Home ML": "-110"
         },
         "Game": {
             "Start Time": "Today \u00b7 4:10 PM",
             "Total": "7.5",
-            "Over Total Odds": "-110",
-            "Under Total Odds": "-110",
+            "Over Total Odds": "-105",
+            "Under Total Odds": "-115",
             "League": "MLB"
         }
     },
@@ -177,9 +177,9 @@
         },
         "Game": {
             "Start Time": "Today \u00b7 4:10 PM",
-            "Total": "8.5",
-            "Over Total Odds": "-130",
-            "Under Total Odds": "+110",
+            "Total": "9.5",
+            "Over Total Odds": "-105",
+            "Under Total Odds": "-115",
             "League": "MLB"
         }
     },
@@ -281,14 +281,14 @@
         "ESPN Away Odds": {
             "Spread": "+1.5",
             "Spread Odds": "-160",
-            "Away ML": "+140"
+            "Away ML": "+135"
         },
         "Home Team": "SEA Mariners",
         "Home TeamID": "300faa7b",
         "ESPN Home Odds": {
             "Spread": "-1.5",
             "Spread Odds": "+135",
-            "Home ML": "-165"
+            "Home ML": "-155"
         },
         "Game": {
             "Start Time": "Today \u00b7 10:10 PM",
