+++ Scrapers/Data/ESPN/NBA.json
@@ -3,15 +3,15 @@
         "Away Team": "BOS Celtics",
         "Away Team Rank Name": "8a2d1e91",
         "ESPN Away Odds": {
-            "Spread": "-16.5",
-            "Spread Odds": "-105",
+            "Spread": "-15.5",
+            "Spread Odds": "-115",
             "Away ML": "-1600"
         },
         "Home Team": "ATL Hawks",
         "Home Team Rank Name": "354ef58",
         "ESPN Home Odds": {
-            "Spread": "+16.5",
-            "Spread Odds": "-115",
+            "Spread": "+15.5",
+            "Spread Odds": "-105",
             "Home ML": "+800"
         },
         "Game": {
