--- Scrapers/Data/mgm.json
@@ -1,1 +1,0 @@
-[]
