--- Scrapers/Data/nbaFD.json
@@ -1,21 +1,0 @@
-{
-    "Away Team": "N/A",
-    "FD Away Odds": {
-        "Spread": "N/A",
-        "Spread Odds": "N/A",
-        "Away ML": "N/A"
-    },
-    "Home Team": "N/A",
-    "FD Home Odds": {
-        "Spread": "N/A",
-        "Spread Odds": "N/A",
-        "Home ML": "N/A"
-    },
-    "Game": {
-        "Start Time": "N/A",
-        "Total": "N/A",
-        "Over Total Odds": "N/A",
-        "Under Total Odds": "N/A",
-        "League": "NBA"
-    }
-}
