+++ Scrapers/Data/not_implemented/cbbdk.json
@@ -1,0 +1,209 @@
+[
+    {
+        "Away Team": "Wagner",
+        "Away Team Rank Name": "Wagner",
+        "DK Away Odds": {
+            "Spread": "+3",
+            "Spread Odds": "\u2212108",
+            "Away ML": "+142"
+        },
+        "Home Team": "Howard",
+        "Home Team Rank Name": "Howard",
+        "DK Home Odds": {
+            "Spread": "-3",
+            "Spread Odds": "-3",
+            "Home ML": "\u2212170"
+        },
+        "Game": {
+            "Start Time": "6:45PM",
+            "Total": "128",
+            "Over Total Odds": "\u2212108",
+            "Under Total Odds": "\u2212112",
+            "League": "NCAAB"
+        }
+    },
+    {
+        "Away Team": "Cornell",
+        "Away Team Rank Name": "Cornell",
+        "DK Away Odds": {
+            "Spread": "+12",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+575"
+        },
+        "Home Team": "Ohio State",
+        "Home Team Rank Name": "Ohio St",
+        "DK Home Odds": {
+            "Spread": "-12",
+            "Spread Odds": "-12",
+            "Home ML": "\u2212750"
+        },
+        "Game": {
+            "Start Time": "7:00PM",
+            "Total": "147.5",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NCAAB"
+        }
+    },
+    {
+        "Away Team": "Boston College",
+        "Away Team Rank Name": "Boston Col",
+        "DK Away Odds": {
+            "Spread": "+5",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+180"
+        },
+        "Home Team": "Providence",
+        "Home Team Rank Name": "Providence",
+        "DK Home Odds": {
+            "Spread": "-5",
+            "Spread Odds": "-5",
+            "Home ML": "\u2212218"
+        },
+        "Game": {
+            "Start Time": "7:00PM",
+            "Total": "137.5",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NCAAB"
+        }
+    },
+    {
+        "Away Team": "North Texas",
+        "Away Team Rank Name": "North Texas",
+        "DK Away Odds": {
+            "Spread": "+2.5",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+130"
+        },
+        "Home Team": "LSU",
+        "Home Team Rank Name": "LSU",
+        "DK Home Odds": {
+            "Spread": "-2.5",
+            "Spread Odds": "-2.5",
+            "Home ML": "\u2212155"
+        },
+        "Game": {
+            "Start Time": "7:00PM",
+            "Total": "132.5",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NCAAB"
+        }
+    },
+    {
+        "Away Team": "Xavier",
+        "Away Team Rank Name": "Xavier",
+        "DK Away Odds": {
+            "Spread": "+1.5",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+110"
+        },
+        "Home Team": "Georgia",
+        "Home Team Rank Name": "Georgia",
+        "DK Home Odds": {
+            "Spread": "-1.5",
+            "Spread Odds": "-1.5",
+            "Home ML": "\u2212130"
+        },
+        "Game": {
+            "Start Time": "7:00PM",
+            "Total": "146.5",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NCAAB"
+        }
+    },
+    {
+        "Away Team": "South Florida",
+        "Away Team Rank Name": "S Florida",
+        "DK Away Odds": {
+            "Spread": "+5",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+185"
+        },
+        "Home Team": "UCF",
+        "Home Team Rank Name": "UCF",
+        "DK Home Odds": {
+            "Spread": "-5",
+            "Spread Odds": "-5",
+            "Home ML": "\u2212225"
+        },
+        "Game": {
+            "Start Time": "9:00PM",
+            "Total": "136.5",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NCAAB"
+        }
+    },
+    {
+        "Away Team": "Minnesota",
+        "Away Team Rank Name": "Minnesota",
+        "DK Away Odds": {
+            "Spread": "+4",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+170"
+        },
+        "Home Team": "Butler",
+        "Home Team Rank Name": "Butler",
+        "DK Home Odds": {
+            "Spread": "-4",
+            "Spread Odds": "-4",
+            "Home ML": "\u2212205"
+        },
+        "Game": {
+            "Start Time": "9:00PM",
+            "Total": "145",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NCAAB"
+        }
+    },
+    {
+        "Away Team": "Kansas State",
+        "Away Team Rank Name": "Kansas St",
+        "DK Away Odds": {
+            "Spread": "+4.5",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+170"
+        },
+        "Home Team": "Iowa",
+        "Home Team Rank Name": "Iowa",
+        "DK Home Odds": {
+            "Spread": "-4.5",
+            "Spread Odds": "-4.5",
+            "Home ML": "\u2212205"
+        },
+        "Game": {
+            "Start Time": "9:00PM",
+            "Total": "152.5",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NCAAB"
+        }
+    },
+    {
+        "Away Team": "UC Irvine",
+        "Away Team Rank Name": "UC Irvine",
+        "DK Away Odds": {
+            "Spread": "+7.5",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+295"
+        },
+        "Home Team": "Utah",
+        "Home Team Rank Name": "Utah",
+        "DK Home Odds": {
+            "Spread": "-7.5",
+            "Spread Odds": "-7.5",
+            "Home ML": "\u2212375"
+        },
+        "Game": {
+            "Start Time": "11:00PM",
+            "Total": "147",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NCAAB"
+        }
+    }
+]
