+++ Scrapers/Data/not_implemented/mgm.json
@@ -1,0 +1,1 @@
+[]
