+++ Scrapers/Data/not_implemented/nbadk.json
@@ -1,0 +1,186 @@
+[
+    {
+        "Away Team": "CLE Cavaliers",
+        "Away Team Rank Name": "Cleveland",
+        "DK Away Odds": {
+            "Spread": "+7",
+            "Spread Odds": "\u2212112",
+            "Away ML": "+220"
+        },
+        "Home Team": "IND Pacers",
+        "Home Team Rank Name": "Indiana",
+        "DK Home Odds": {
+            "Spread": "-7",
+            "Spread Odds": "\u2212108",
+            "Home ML": "\u2212270"
+        },
+        "Game": {
+            "Start Time": "7:10PM",
+            "Total": "224.5",
+            "Over Total Odds": "\u2212112",
+            "Under Total Odds": "\u2212108",
+            "League": "NBA"
+        }
+    },
+    {
+        "Away Team": "DET Pistons",
+        "Away Team Rank Name": "Detroit",
+        "DK Away Odds": {
+            "Spread": "+16",
+            "Spread Odds": "\u2212108",
+            "Away ML": "+850"
+        },
+        "Home Team": "BOS Celtics",
+        "Home Team Rank Name": "Boston",
+        "DK Home Odds": {
+            "Spread": "-16",
+            "Spread Odds": "\u2212112",
+            "Home ML": "\u22121450"
+        },
+        "Game": {
+            "Start Time": "7:40PM",
+            "Total": "224",
+            "Over Total Odds": "\u2212110",
+            "Under Total Odds": "\u2212110",
+            "League": "NBA"
+        }
+    },
+    {
+        "Away Team": "MIA Heat",
+        "Away Team Rank Name": "Miami",
+        "DK Away Odds": {
+            "Spread": "+2.5",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+114"
+        },
+        "Home Team": "PHI 76ers",
+        "Home Team Rank Name": "Philadelphia",
+        "DK Home Odds": {
+            "Spread": "-2.5",
+            "Spread Odds": "\u2212110",
+            "Home ML": "\u2212135"
+        },
+        "Game": {
+            "Start Time": "7:40PM",
+            "Total": "208.5",
+            "Over Total Odds": "\u2212112",
+            "Under Total Odds": "\u2212108",
+            "League": "NBA"
+        }
+    },
+    {
+        "Away Team": "POR Trail Blazers",
+        "Away Team Rank Name": "Portland",
+        "DK Away Odds": {
+            "Spread": "+8",
+            "Spread Odds": "\u2212110",
+            "Away ML": "+260"
+        },
+        "Home Team": "CHI Bulls",
+        "Home Team Rank Name": "Chicago",
+        "DK Home Odds": {
+            "Spread": "-8",
+            "Spread Odds": "\u2212110",
+            "Home ML": "\u2212325"
+        },
+        "Game": {
+            "Start Time": "8:10PM",
+            "Total": "214",
+            "Over Total Odds": "\u2212108",
+            "Under Total Odds": "\u2212112",
+            "League": "NBA"
+        }
+    },
+    {
+        "Away Team": "MIN Timberwolves",
+        "Away Team Rank Name": "Minnesota",
+        "DK Away Odds": {
+            "Spread": "-6.5",
+            "Spread Odds": "\u2212105",
+            "Away ML": "\u2212245"
+        },
+        "Home Team": "UTA Jazz",
+        "Home Team Rank Name": "Utah",
+        "DK Home Odds": {
+            "Spread": "+6.5",
+            "Spread Odds": "\u2212115",
+            "Home ML": "+200"
+        },
+        "Game": {
+            "Start Time": "9:10PM",
+            "Total": "223",
+            "Over Total Odds": "\u2212108",
+            "Under Total Odds": "\u2212112",
+            "League": "NBA"
+        }
+    },
+    {
+        "Away Team": "NY Knicks",
+        "Away Team Rank Name": "New York",
+        "DK Away Odds": {
+            "Spread": "+5",
+            "Spread Odds": "\u2212112",
+            "Away ML": "+164"
+        },
+        "Home Team": "GS Warriors",
+        "Home Team Rank Name": "Golden State",
+        "DK Home Odds": {
+            "Spread": "-5",
+            "Spread Odds": "\u2212108",
+            "Home ML": "\u2212198"
+        },
+        "Game": {
+            "Start Time": "10:10PM",
+            "Total": "211",
+            "Over Total Odds": "\u2212112",
+            "Under Total Odds": "\u2212108",
+            "League": "NBA"
+        }
+    },
+    {
+        "Away Team": "MEM Grizzlies",
+        "Away Team Rank Name": "Memphis",
+        "DK Away Odds": {
+            "Spread": "+10",
+            "Spread Odds": "\u2212108",
+            "Away ML": "+360"
+        },
+        "Home Team": "SAC Kings",
+        "Home Team Rank Name": "Sacramento",
+        "DK Home Odds": {
+            "Spread": "-10",
+            "Spread Odds": "\u2212112",
+            "Home ML": "\u2212470"
+        },
+        "Game": {
+            "Start Time": "10:10PM",
+            "Total": "223.5",
+            "Over Total Odds": "\u2212108",
+            "Under Total Odds": "\u2212112",
+            "League": "NBA"
+        }
+    },
+    {
+        "Away Team": "ATL Hawks",
+        "Away Team Rank Name": "Atlanta",
+        "DK Away Odds": {
+            "Spread": "+8.5",
+            "Spread Odds": "\u2212112",
+            "Away ML": "+285"
+        },
+        "Home Team": "LA Lakers",
+        "Home Team Rank Name": "LA Lakers",
+        "DK Home Odds": {
+            "Spread": "-8.5",
+            "Spread Odds": "\u2212108",
+            "Home ML": "\u2212360"
+        },
+        "Game": {
+            "Start Time": "10:40PM",
+            "Total": "226",
+            "Over Total Odds": "\u2212112",
+            "Under Total Odds": "\u2212108",
+            "League": "NBA"
+        }
+    }
+]
